% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{agg_sleep_weekday}
\alias{agg_sleep_weekday}
\title{A function to preprocess sleep data for the Person object.}
\usage{
agg_sleep_weekday(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
A tidy data frame with the columns weekday, measure, and hours
}
\description{
Preprocesses data to be used by the plot_sleep_weekday() 
function.  Specifically, it calculates the sleep duration and time asleep for 
each day of the week (in hours).
}
\examples{
data(EX)
agg_sleep_weekday(person = EX)

}
