% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EX.R
\docType{data}
\name{EX}
\alias{EX}
\title{A subset of the data for one user for about one month, from 2017-01-19 to
2017-02-17, containing \code{fitbit_daily}, \code{fitbit_intraday}, and 
\code{util} data
frames.}
\format{An object of class \code{Person} (inherits from \code{R6}) of length 14.}
\usage{
EX
}
\description{
A subset of the data for one user for about one month, from 2017-01-19 to
2017-02-17, containing \code{fitbit_daily}, \code{fitbit_intraday}, and 
\code{util} data
frames.
}
\keyword{data}
