% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelogr.R
\docType{package}
\name{lifelogr}
\alias{lifelogr}
\alias{lifelogr-package}
\title{lifelogr: A package for life logging in R.}
\description{
The lifelogr package provides a framework for creating visualizations and
experimenting on onself.  It provides an example dataset: \code{EX}.
}
\section{Foo functions}{

The foo functions ...
}

\section{To learn more about lifelogr, start with the vignette}{
 
`browseVignettes(package = "lifelogr")`
}

