% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{lifelogrApp}
\alias{lifelogrApp}
\title{Run the Shiny app.}
\usage{
lifelogrApp()
}
\description{
Meant to be used as an example showcasing the visualizations.
Uses the EX Person instance used throughout this package.
}
\examples{
lifelogrApp
}
