% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person.R
\docType{class}
\name{Person}
\alias{Person}
\title{A \code{Person} object is a complete view of an individual over a certain
time period, as seen through data from multiple sources}
\format{An \code{\link{R6Class}} generator object}
\usage{
Person
}
\description{
\code{Person} is an object that encapsulates an individual's 
data over a specified date range (start and end date stored as \code{Date} 
objects.
An individual consists of basic information, such as name, age,
and gender (a \code{list} with named elements), data from their self-tracking 
devices such as Fitbit, Apple health, etc. (data from each source is a tibble
dataframe), individual goals such as target steps (\code{numeric}), 
additional data from self-tracking apps or one's own collection system 
(stored as a tibble dataframe), and ways of grouping the data a user may be 
interested in, such as grouping by seasons, or comparing weekend to weekday 
behavior and health (stored as a list of named dataframes, which each contain 
group assignments).
}
\section{Fields}{

\describe{
\item{\code{fitbit_daily}}{tibble dataframe of fitbit variables (for user account
info provided) observed daily}

\item{\code{fitbit_intraday}}{tibble dataframe of fitbit variables (for user account
info provided) observed multiple times a day}

\item{\code{util}}{tibble dataframe that maps each date in the date range to utility
information about that date}

\item{\code{target_steps}}{the person's target number of steps (numeric) for each
day (default 10,000)}

\item{\code{start_date}}{start of user's date range of interest (Date object)}

\item{\code{end_date}}{end of user's date range of interest (Date object)}

\item{\code{user_info}}{provided user info, such as "age", "gender", "name" (list)}

\item{\code{groupings}}{named list of dataframes, each with two columns - a known 
variable, and group, with the group assignment for observations where 
that variable has appropriate value}

\item{\code{apple}}{tibble dataframe of user's provided Apple Health data}

\item{\code{addl_data}}{dataframe of data from another source provided by user}

\item{\code{addl_data2}}{dataframe of data from another source provided by user}
}}

\section{Methods}{


\describe{
  \item{\code{Person$new(fitbit_user_email, fitbit_user_pw, user_info = NA,
                          apple_data_file, target_steps, addl_data,
                          addl_data2, group_assignments, start_date, 
                          end_date)}}{Creates a new \code{Person}
        with specified data, and data from provided Fitbit account. If 
        provided, start_date and end_date must be characters with "%Y-%m-%d". 
        All defaults are \code{NA} - user can provide sources of data
        of interest.}}
}

\examples{
library("lifelogr")
data(EX)

group_months <- data.frame("month" = c("Jan", "Feb", "Mar", "Apr", "May",
                                       "Jun", "Jul", "Aug",
                                       "Sep", "Oct", "Nov", "Dec"),
                                       "group" = c(0, 0, 0, 1, 1, 1, 1, 1, 
                                                   1, 0, 0, 0))
ash <- Person$new(user_info = list("name" = "Ash", "age" = 26,
                    "gender" = "female"), 
                    target_steps = 20000,
                    addl_data = EX,
                    group_assignments = list("group_months" = group_months),
                    start_date = "2017-03-11",
                    end_date = "2017-03-12")

\dontrun{                
bailey <- Person$new(fitbit_user_email = "bailey@gmail.com",
                 fitbit_user_pw = "baileypw",
                 #apple_data_file = "apple.csv",
                 start_date = "2017-03-11",
                 end_date = "2017-03-12")}
                 
}
\keyword{datasets}
