% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_cal}
\alias{plot_cal}
\title{Plot calories over time.}
\usage{
plot_cal(Person)
}
\arguments{
\item{Person}{The user's data}
}
\value{
NULL, but plot printed to screen
}
\description{
Prints a line plot plotting calories burned over time.  If
    calories consumed are in the dataset, it also plots calories consumed.
}
\examples{
data(EX)
plot_cal(EX)
}
