% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_d}
\alias{plot_d}
\title{Line graph for continuous variable(s).}
\usage{
plot_d(person, measures)
}
\arguments{
\item{person}{An instance of the Person class}

\item{measures}{A character vector of length one or more indicating the 
variable(s) of interest.  Options include: "steps", "floors", "distance",
 "calories", "mins_very", "rest_hr".}
}
\value{
NULL, but plot printed to screen
}
\description{
A "quick-and-dirty" approach to plotting a generic line graph 
    with default axis labels.  Can plot one or more variables.
}
\examples{
data(EX)
plot_d(EX, "steps")
plot_d(EX, c("steps", "distance"))
}
