% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_daily}
\alias{plot_daily}
\title{Plot daily health totals.}
\usage{
plot_daily(person, measure_var = "all", ...)
}
\arguments{
\item{person}{An instance of the Person class}

\item{measure_var}{Default is to print all six plots.  Options include:
"steps", "floors", "distance", "calories", "mins_very", "rest_hr", "all".}

\item{...}{Extra arguments used to specify unit for the distance plot.}
}
\value{
NULL, but plots printed to screen
}
\description{
Prints one of six plots, each showing daily totals over time.
}
\examples{
data(EX)
plot_daily(EX, "steps")
plot_daily(EX, "distance", "km")

}
