% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_daily_all}
\alias{plot_daily_all}
\title{Plot a series of six graphs.}
\usage{
plot_daily_all(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plots printed to screen
}
\description{
Prints six plots, each showing daily totals over time: 
1.  Steps
2.  Floors
3.  Distance: in the default unit, miles
4.  Calories
5.  Minutes 'very active'
6.  Resting heart rate
}
\examples{
data(EX)
plot_daily_all(EX)

}
