% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_distance}
\alias{plot_distance}
\title{Plot distance per day over time.}
\usage{
plot_distance(person, unit = "mi")
}
\arguments{
\item{person}{An instance of the Person class}

\item{unit}{a unit of distance, 'mi' or 'km'.  The default value is 'mi'}
}
\value{
NULL, but plot printed to screen
}
\description{
Prints a line plot plotting distance in miles or kilometers per 
    day over time.
}
\examples{
data(EX)
plot_distance(EX)
plot_distance(EX, "mi")
plot_distance(EX, "km")
}
