% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_floors}
\alias{plot_floors}
\title{Plot number of floors per day over time.}
\usage{
plot_floors(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plot printed to screen
}
\description{
Prints a line plot plotting number of floors per day over time.
}
\examples{
data(EX)
plot_floors(EX)
}
