% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_intraday.R
\name{plot_intraday_all}
\alias{plot_intraday_all}
\title{Plot all intraday variables.}
\usage{
plot_intraday_all(Person)
}
\arguments{
\item{Person}{The user's data.}
}
\value{
NULL, plots print to screen
}
\description{
Plots all seven intraday variables using default settings.
}
\examples{
data(EX)
plot_intraday_all(EX)
}
\seealso{
\code{\link{plot_i}}
}
