% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_mins_very}
\alias{plot_mins_very}
\title{Plot minutes 'very active' over time.}
\usage{
plot_mins_very(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plot printed to screen
}
\description{
Prints a line plot plotting minutes 'very active' per day over 
    time.  'Very active' is a subjective term defined by fitbit.
}
\examples{
data(EX)
plot_mins_very(EX)
}
