% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_rest_hr}
\alias{plot_rest_hr}
\title{Plot resting heart rate over time.}
\usage{
plot_rest_hr(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plot printed to screen
}
\description{
Prints a line plot plotting heart rate (in beats per minute)
    over time.  According to the National Institute of Health, the average
    resting heart rate for persons 10 and older (including seniors) is 60 - 
    100.  However, well-trained athletes can have resting heart rates between
     40 and 60.
}
\examples{
data(EX)
plot_rest_hr(EX)
}
\seealso{
\url{http://www.heart.org/HEARTORG/HealthyLiving/PhysicalActivity/FitnessBasics/Target-Heart-Rates_UCM_434341_Article.jsp#.WM3bCxiZMdU}
}
