% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep}
\alias{plot_sleep}
\title{Plot sleep.}
\usage{
plot_sleep(Person, plot_type = "all")
}
\arguments{
\item{Person}{The user's data}

\item{plot_type}{The type of plot.  Options include: "by_weekday", 
"by_start_end_time", "by_datetime", "by_restless_prop", "by_restless_min", 
"by_quality".  Default is to plot all six.}
}
\value{
A ggplot2 object
}
\description{
Prints one of six plots: two are related to quantity of sleep, 
and four are related to quality of sleep
1.  Sleep by day of week (bar graph)
2.  Start and end of sleep period for each day in the range
3.  Duration of sleep and time asleep over time
4.  Proportion of time spent restless out of total sleep duration over time
5.  Time spent restless over time (in minutes)
6.  Sleep quality over time (subjective score, out of 100)
}
\examples{
data(EX)
plot_sleep(EX)

}
