% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_all}
\alias{plot_sleep_all}
\title{Plot a series of six sleep graphs.}
\usage{
plot_sleep_all(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plot prints to screen
}
\description{
Prints six plots: two are related to quantity of sleep, and 
four are related to quality of sleep
1.  Sleep by day of week (bar graph)
2.  Start and end of sleep period for each day in the range
3.  Duration of sleep and time asleep over time
4.  Proportion of time spent restless out of total sleep duration over time
5.  Time spent restless over time (in minutes)
6.  Sleep quality over time (subjective score, out of 100)
}
\examples{
data(EX)
plot_sleep_all(person = EX)

}
