% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_over_time}
\alias{plot_sleep_over_time}
\title{A function to plot sleep over time.}
\usage{
plot_sleep_over_time(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plots print to screen
}
\description{
Returns a line plot plotting sleep over time.  Includes sleep
    duration and time asleep (in hours).
}
\examples{
data(EX)
plot_sleep_over_time(person = EX)

}
