% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_quality}
\alias{plot_sleep_quality}
\title{A function to plot sleep quality over time.}
\usage{
plot_sleep_quality(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plots print to screen
}
\description{
Returns a line plot plotting sleep quality over time.  Sleep
quality is a subjective score given by Fitbit
}
\examples{
data(EX)
plot_sleep_quality(person = EX)

}
