% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_restless_min}
\alias{plot_sleep_restless_min}
\title{A function to plot the minutes of restless sleep over time.}
\usage{
plot_sleep_restless_min(Person)
}
\arguments{
\item{Person}{The user's data}
}
\value{
A ggplot2 object, prints to screen
}
\description{
Returns a line plot plotting the length of restless sleep 
over time (in minutes).
}
\examples{
data(EX)
plot_sleep_restless_min(EX)

}
