% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_weekday}
\alias{plot_sleep_weekday}
\title{A function to plot sleep by day of week.}
\usage{
plot_sleep_weekday(Person)
}
\arguments{
\item{Person}{The user's data}
}
\value{
A ggplot2 object, prints to screen
}
\description{
Returns a bar graph plotting sleep by day of week (Sunday, 
Monday, ...).
}
\examples{
data(EX)
plot_sleep_weekday(EX)

}
