% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{plot_steps}
\alias{plot_steps}
\title{Plot steps per day over time.}
\usage{
plot_steps(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plot printed to screen
}
\description{
Prints a line plot plotting steps per day over time.  The
    reference line refers to the user's target number of steps.
}
\examples{
data(EX)
plot_steps(EX)

}
