% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_daily.R
\name{tidy_multi_meas_data}
\alias{tidy_multi_meas_data}
\title{Tidy daily data.}
\usage{
tidy_multi_meas_data(data)
}
\arguments{
\item{data}{Data frame or tibble with a column named 'date' and other columns
of interest.}
}
\value{
Tidy tibble with the columns date, measures, and value.
}
\description{
Tidy daily data with multiple measures.
}
\examples{
a <- tibble::tibble(date = 
         lubridate::ymd("1970-01-01", "1970-01-02", "1970-01-03"), 
         sleepDurationHrs = c(7.5, 8.0, 7.9), 
         minAsleepHrs = c(7.4, 7.0, 7.7))
tidy_multi_meas_data(a)
}
