% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifelogr.R
\docType{package}
\name{lifelogr}
\alias{lifelogr}
\alias{lifelogr-package}
\title{lifelogr: A package for life logging in R.}
\description{
The lifelogr package provides a framework for creating visualizations 
and experimenting on onself using self-tracking health data from multiple
sources. It provides an example of what a user's combined dataset 
might look like: \code{EX}.
}
\details{
To learn more about lifelogr, start with the vignette: 
\code{browseVignettes(package = "lifelogr")}
}
