% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_intraday.R
\name{plot_intraday_all}
\alias{plot_intraday_all}
\title{Plot all intraday variables.}
\usage{
plot_intraday_all(person, avg_to_get_typical_day = TRUE)
}
\arguments{
\item{person}{An instance of the Person class.}

\item{avg_to_get_typical_day}{Logical vector of length 1.  If TRUE, plot 
gives an aggregate of the variable over the course of a typical day.  If 
FALSE, plot gives the variable at every interval over the range specified
 when the Person object was instantiated.}
}
\value{
NULL, plots print to screen
}
\description{
Plots all seven intraday variables using default settings.
}
\examples{
data(EX)
plot_intraday_all(EX)
}
\seealso{
\code{\link{plot_i}}
}
