% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_restless_min}
\alias{plot_sleep_restless_min}
\title{A function to plot the minutes of restless sleep over time.}
\usage{
plot_sleep_restless_min(person)
}
\arguments{
\item{person}{An instance of the Person class}
}
\value{
NULL, but plots print to screen
}
\description{
Returns a line plot plotting the length of restless sleep 
    over time (in minutes).
}
\examples{
data(EX)
plot_sleep_restless_min(person = EX)

}
