% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_sleep.R
\name{plot_sleep_start_end}
\alias{plot_sleep_start_end}
\title{A function to plot sleep each night by start time and end time.}
\usage{
plot_sleep_start_end(person, color_var = "day_type")
}
\arguments{
\item{person}{An instance of the Person class}

\item{color_var}{"day_type" by default for weekend/weekday, or "day_of_week"
for day of week.  Determines color of the lines.}
}
\value{
NULL, but plots print to screen
}
\description{
Returns a plot with start time of sleep and end time of sleep
    each night, colored by weekday vs. weekend.
}
\examples{
data(EX)
plot_sleep_start_end(person = EX)
plot_sleep_start_end(person = EX, "day_of_week")

}
